requirements.txt
fastapi~=0.115.12
uvicorn[standard]
idna~=3.7
cryptography~=44.0.1
pip~=25.1
attrs~=23.2.0
distro~=1.9.0
wheel~=0.42.0
pillow~=10.2.0
Jinja2~=3.1.2
filelock~=3.13.1
rich~=13.7.1
Pygments~=2.17.2
markdown-it-py~=3.0.0
PyYAML~=6.0.1
click~=8.1.6
mdurl~=0.1.2
MarkupSafe~=2.1.5
certifi~=2023.11.17
requests~=2.31.0
psutil~=5.9.8
SQLAlchemy~=2.0.41
pydantic~=2.11.5
passlib~=1.7.4
pydantic[email]~=2.11
pydantic-settings~=2.9.1
psycopg2-binary
PyJWT>=2.0.0
python-multipart
jwt~=1.3.1
pytest~=8.3.5
python-jose~=3.4.0